instances() shows Placement.AvailabilityZone
After the great AZ outage of 2016 and with increased popularity
of iterating across Regions in a for loop, I've overcome my
hatred of long output and added AZ to instance output.

In future I hope to reduce the length of the standard output
to reduce the offence to right thinking users. Until then, I
trust the tradeoff will be tolerated and perhaps found useful.